tired of random.randint
this will use the date(%d) and Hour(%H) as to choose the GIF. Thus it is expected to cover all the gifs under a particular key in 24 hours. 
Lets see if it works or not. 
I also have to delete some of Makarov's unnecessary quotes as their cannot be more than 24 quotes in a key.